requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy==2.0.22
pymysql==1.1.0
pydantic>=1.8.0
pydantic-settings>=2.0.3
minio==7.1.16
psutil==5.9.5
python-multipart>=0.0.5
cryptography==41.0.5
alembic==1.12.0
tortoise-orm>=0.17.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
aiofiles>=0.7.0
python-magic>=0.4.24
Pillow>=8.3.1
pillow-heif>=0.4.0
asyncpg>=0.24.0 